Moved some code around to prevent a cyclic import
