+++ pypopquiz/popquiz.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import pypopquiz as ppq
 import pypopquiz.io
 import pypopquiz.sheets
+import pypopquiz.sources
 import pypopquiz.video
 
 
@@ -32,7 +33,7 @@ def popquiz(input_file: Path, output_dir: Path, backend: str) -> None:
 
     for question in input_data["questions"]:
         for source in question["sources"]:
-            ppq.io.get_source(source, output_dir, input_file.parent)
+            ppq.sources.get_source(source, output_dir, input_file.parent)
 
     q_videos, a_videos = [], []
     for index, question in enumerate(input_data["questions"]):
